// frontend/src/generated/models/performanceWorkspaceResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * CFO Command Center API
 * OpenAPI spec version: 0.3.0
 */
import type { PerformanceWorkspaceResponseAnomaliesItem } from './performanceWorkspaceResponseAnomaliesItem';
import type { PerformanceWorkspaceResponseAssurance } from './performanceWorkspaceResponseAssurance';
import type { PerformanceWorkspaceResponseCommentaryRequirementsItem } from './performanceWorkspaceResponseCommentaryRequirementsItem';
import type { WorkspaceContextResponse } from './workspaceContextResponse';
import type { PerformanceWorkspaceResponseKpiTreeItem } from './performanceWorkspaceResponseKpiTreeItem';
import type { PerformanceWorkspaceResponseLineage } from './performanceWorkspaceResponseLineage';
import type { PerformanceWorkspaceResponseMetricsItem } from './performanceWorkspaceResponseMetricsItem';
import type { PerformanceWorkspaceResponseTrendItem } from './performanceWorkspaceResponseTrendItem';
import type { PerformanceWorkspaceResponseVarianceBridge } from './performanceWorkspaceResponseVarianceBridge';

export interface PerformanceWorkspaceResponse {
  anomalies?: PerformanceWorkspaceResponseAnomaliesItem[];
  as_of: string;
  assurance?: PerformanceWorkspaceResponseAssurance;
  commentary_requirements?: PerformanceWorkspaceResponseCommentaryRequirementsItem[];
  context: WorkspaceContextResponse;
  kpi_tree?: PerformanceWorkspaceResponseKpiTreeItem[];
  lineage?: PerformanceWorkspaceResponseLineage;
  metrics?: PerformanceWorkspaceResponseMetricsItem[];
  projection_version: number;
  source_snapshot_ids?: string[];
  trend?: PerformanceWorkspaceResponseTrendItem[];
  variance_bridge?: PerformanceWorkspaceResponseVarianceBridge;
}